Update gemspec version
With the version name that was in place it was considered a prerelease

Many gem developers have versions of their gem ready to go out for testing or “edge” releases before a big gem release. RubyGems supports the concept of “prerelease” versions, which could be betas, alphas, or anything else that isn’t ready as a regular release.

Taking advantage of this is easy. All you need is one or more letters in the gem version.[1]

[1] https://guides.rubygems.org/patterns/